\name{liveNiLeDAM}
\alias{liveNiLeDAM}

\title{Graphical Web User Interface for NiLeDAM}
\description{Starts the NiLeDAM GUI}
\usage{
liveNiLeDAM()
}

\arguments{
  No arguments.
}

\value{This function starts the graphical user interface with the default 
system browser. This interface is more lickely to work properly with Firefox
\url{http://www.mozilla.org/fr/firefox/new/}. In case Firefox is not your 
default browser, copy/paste http://localhost:8100 into the address bar.
}

\author{
Nathalie Villa-Vialaneix \email{nathalie.villa@toulouse.inra.fr}
}

