% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveNILEDAM.R
\name{liveNiLeDAM}
\alias{liveNiLeDAM}
\title{Graphical Web User Interface for 'NiLeDAM'}
\usage{
liveNiLeDAM()
}
\value{
Starts the 'NiLeDAM' GUI.
}
\description{
This function starts the graphical user interface with the default 
system browser. This interface is more likely to work properly with Firefox
\url{https://www.mozilla.org/fr/firefox/new/}. In case Firefox is not your 
default browser, copy/paste http://localhost:8100 into the address bar.
}
\author{
Aurélie Mercadié, \email{aurelie.mercadie@inrae.fr}

Nathalie Vialaneix, \email{nathalie.vialaneix@inrae.fr}
}
